requirements.txt
requests
beautifulsoup4
newspaper3k
rake-nltk
readability-lxml
nltk
